+++ contracts/src/PairOpenSumSwap.sol
@@ -3,48 +3,39 @@ pragma solidity 0.6.8;
 
 import "@openzeppelin/contracts/math/SafeMath.sol";
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-import "../interfaces/stableswap/IOpenSumSwap.sol";
-import "./ISwappaPairV1.sol";
+import "./interfaces/stableswap/IOpenSumSwap.sol";
+import "./interfaces/ISwappaPairV1.sol";
 
 contract PairOpenSumSwap is ISwappaPairV1 {
-	using SafeMath for uint;
+  using SafeMath for uint256;
 
-	function swap(
-		address input,
-		address output,
-		address to,
-		bytes calldata data
-	) external override {
-		address swapPoolAddr = parseData(data);
-		uint inputAmount = ERC20(input).balanceOf(address(this));
-		require(
-			ERC20(input).approve(swapPoolAddr, inputAmount),
-			"PairOpenSumSwap: approve failed!");
-		uint outputAmount = IOpenSumSwap(swapPoolAddr).swap(
-			input,
-			output,
-			inputAmount,
-			inputAmount,
-			block.timestamp);
-		require(
-			ERC20(output).transfer(to, outputAmount),
-			"PairOpenSumSwap: transfer failed!");
-	}
+  function swap(
+    address input,
+    address output,
+    address to,
+    bytes calldata data
+  ) external override {
+    address swapPoolAddr = parseData(data);
+    uint256 inputAmount = ERC20(input).balanceOf(address(this));
+    require(ERC20(input).approve(swapPoolAddr, inputAmount), "PairOpenSumSwap: approve failed!");
+    uint256 outputAmount = IOpenSumSwap(swapPoolAddr).swap(input, output, inputAmount, inputAmount, block.timestamp);
+    require(ERC20(output).transfer(to, outputAmount), "PairOpenSumSwap: transfer failed!");
+  }
 
-	function parseData(bytes memory data) private pure returns (address swapPoolAddr) {
-		require(data.length == 20, "PairOpenSumSwap: invalid data!");
-		assembly {
-			swapPoolAddr := mload(add(data, 20))
-		}
-	}
+  function parseData(bytes memory data) private pure returns (address swapPoolAddr) {
+    require(data.length == 20, "PairOpenSumSwap: invalid data!");
+    assembly {
+      swapPoolAddr := mload(add(data, 20))
+    }
+  }
 
-	function getOutputAmount(
-		address input,
-		address output,
-		uint amountIn,
-		bytes calldata data
-	) external view override returns (uint amountOut) {
-		// no fees are taken
-		return amountIn;
-	}
+  function getOutputAmount(
+    address input,
+    address output,
+    uint256 amountIn,
+    bytes calldata data
+  ) external view override returns (uint256 amountOut) {
+    // no fees are taken
+    return amountIn;
+  }
 }
